3.SDT/1.database-management/python-connect-with-mysql/requirements.txt
mysql-connector-python==8.0.20
protobuf==6.31.1
